Remove the PlacementNew class wrapper
This should help with debug an awful lot.
