+++ Source/Barrage/Public/BarrageDispatch.h
@@ -15,11 +15,6 @@
 #include "Containers/Deque.h"
 #include "BarrageDispatch.generated.h"
 
-struct FBInputPlacementNew
-{
-	char block[48];
-};
-
 enum LayersMap
 {
 	 NON_MOVING = 0,
@@ -55,14 +50,7 @@ class BARRAGE_API UBarrageDispatch : public UTickableWorldSubsystem
 	static inline constexpr float TickRateInDelta = 1.0 / 120.0;
 	
 public:
-	
-	typedef TCircularQueue<FBInputPlacementNew> ThreadFeed;
-	
-	struct FeedMap
-	{
-		std::thread::id That = std::thread::id();
-		TSharedPtr<ThreadFeed> Queue = nullptr;
-	};
+
 	struct TransformUpdate
 	{
 		ObjectKey ObjectKey;
@@ -74,21 +62,12 @@ public:
 	uint8 ThreadAccTicker = 0;
 	typedef TCircularQueue<TransformUpdate> TransformUpdatesForGameThread;
 	TSharedPtr<TransformUpdatesForGameThread> GameTransformPump;
-	FeedMap ThreadAcc[ALLOWED_THREADS_FOR_BARRAGE_PHYSICS];
 	 //this value indicates you have none.
 	mutable FCriticalSection GrowOnlyAccLock;
 
 	// Why would I do it this way? It's fast and easy to debug, and we will probably need to force a thread
 	// order for determinism. this ensures there's a call point where we can institute that.
-	void GrantFeed()
-	{
-	FScopeLock GrantFeedLock(&GrowOnlyAccLock);
-		ThreadAcc[ThreadAccTicker].That = std::this_thread::get_id();
-		MyBARRAGEIndex = ThreadAccTicker;
-		//TODO: expand if we need for rollback powers. could be sliiiick
-		ThreadAcc[ThreadAccTicker].Queue = MakeShareable(new ThreadFeed(1024));
-		++ThreadAccTicker;
-	}
+	void GrantFeed();
 	UBarrageDispatch();
 	
 	virtual ~UBarrageDispatch() override;
